Fixed Docker container, broken Spring routing
The previously closed issue #28 broke some of the routes of the app. This commit
fixes those routes.

Related work items: #28
