+++ src/main/app/src/App.js
@@ -6,7 +6,6 @@ function App() {
          <NavBar/>
       </div>
   );
-  
 }
 
 export default App;
