--- src/main/java/ca/serum390/godzilla/configuration/WebConfig.java
@@ -1,18 +1,0 @@
-package ca.serum390.godzilla.configuration;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.config.EnableWebFlux;
-import org.springframework.web.reactive.config.ResourceHandlerRegistry;
-import org.springframework.web.reactive.config.WebFluxConfigurer;
-
-
-@Configuration
-@EnableWebFlux
-public class WebConfig implements WebFluxConfigurer {
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/resources/**")
-                .addResourceLocations("/public", "classpath:/static/");
-    }
-}
